refactor: what the pen is set to leaves App as a hook
I measured seam cost across what is left. The drawing code is the biggest block
in App - startDraw, onDraw and stopDraw are 343 lines between them - and it
cannot leave: it reads fifty-three App names it does not own, because drawing
genuinely touches everything.

What it *reads* is the opposite shape. Twenty-eight names for which tool, what
colour, how wide and how hard, and the whole group reads four things it does not
own, three of which exist for one guard. Read from everywhere, written from
almost nowhere.

`useToolSettings` takes them, including the two derived values that were the
reason half of this was worth grouping at all: `etool`, since Ruler and Air are
each two tools behind one button, and `toolSize`, since the eraser keeps its own
width so switching to it and back does not lose the size you were drawing with.

`handleSetTool`'s two guards are injected rather than moved. `busy()` is the one
that refuses a tool change while a selection floats or a text is open. The curve
one keeps its "is there one" test on App's side deliberately: `commitCurve` with
nothing to commit still clears the live overlay, and calling it on every tool
switch would wipe a stroke in progress.

App.jsx: 3690 -> 3657 lines, and thirty names out of its namespace: 385 -> 356
reachable App-level names, 80 pieces of state -> 71.

The new import gate caught the two imports this orphaned before I did.
